Need source in Gemfile
Without it when you try to install the gems you get

Your Gemfile has no gem server sources. If you need gems that are not
already on your machine, add a line like this to your Gemfile:
source 'https://rubygems.org'
Could not find gem 'cucumber (~> 3.1.0)' in any of the gem sources
listed in your Gemfile.