Delete contributing.md
I might be wrong, but this file looks like it was imported from another repo - it's a document about contributing to the cucumber project overall rather than a document about contributing to this project. We have decent info about how to contribute in the README already, and I think that info about contributing to Cucumber does belong in the docs somewhere, but we can add that in later.