Update browser-automation.md (#206)
* Update browser-automation.md

test in the example does not pass with an exception:
cucumber.runtime.CucumberException: Arity mismatch: Step Definition 'geoorg.Stepdefs.checkTitle() in file:.../cucumber-selenium-java8/target/test-classes/' with pattern [^the page title should start with "(.*)"$] is declared with 0 parameters. However, the gherkin step has 1 arguments [cheese].

* Update browser-automation.md

'10' is not a character in java 8. changed to long

* Fix typo