Update 10-minute-tutorial.md
When I follow the tutorial and get to the section on creating an outline and providing examples, I get a couple of syntax errors that can be easily remedied with a few changes to where you put quotation marks in the .feature file. Instead of putting quotation marks around every argument in the examples table, put quotation marks around the parameterized terms in the 'Given' and 'Then' statements. This takes care of:
1. IntelliJ telling me I need to implement step definitions for what is technically a different format than what is provided in the doc already, and
2. Spellchecker telling me there's a spelling issue with each of the items in the examples table.