Add "Myth" to each myth title
There's so much unfocused effort that pulls incorrect conclusions from info such as this document, that it seems that these myths are being treated as the opposite. 

Take a look at this page with the mindset that you're new and just did a web search, and didn't read the entire page. With this mindset as a given, it is easy to see how these items can be taken with incorrect, opposite understanding.

Adding "Myth" to the the title of each myth reduces the chance that these items will be taken in a way that is opposite the intent.